Define mock_s3_bucket in conftest.py
